Removed last bits of styled components and replaced to stock CSS from CRA
