+++ src/index.css
@@ -1,6 +1,9 @@
-html, #root, body {
+html,
+#root,
+body {
   height: 100%;
 }
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
@@ -16,6 +19,10 @@ code {
     monospace;
 }
 
+.inline {
+  display: inline;
+}
+
 .flexContainer {
   display: flex;
   align-items: center;
@@ -24,8 +31,21 @@ code {
 .textAlignLeft {
   text-align: left;
 }
+
 .textAlignCenter {
   text-align: center;
+}
+
+.noMargins {
+  margin: 0;
+}
+
+.noMarginTop {
+  margin-top: 0;
+}
+
+.noMarginBottom {
+  margin-bottom: 0;
 }
 
 a:visited {
