+++ src/modals/modals.module.css
@@ -1,0 +1,43 @@
+.flexDiv {
+    height: 100%;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.errorDiv {
+    text-align: center;
+    background: white;
+    color: black;
+    padding: 10px;
+    border-radius: 10px;
+}
+
+.warningDiv {
+    text-align: center;
+    background: white;
+    color: black;
+    padding: 10px;
+    border-radius: 10px;
+}
+
+@keyframes fadeIn {
+    from {
+        opacity: 0
+    }
+
+    to {
+        opacity: 1
+    }
+}
+
+.modalsDiv {
+    position: absolute;
+    height: 100%;
+    width: 100%;
+    top: 0;
+    left: 0;
+
+    animation: fadeIn 0.2s ease-in-out;
+}
