+++ src/routes/AuthRedirect.jsx
@@ -23,6 +23,6 @@ export class AuthRedirect extends React.Component {
   }
   render() {
     // TODO: currently an empty div but might change to a redirect notification or something
-    return <div></div>;
+    return null
   }
 }
