+++ src/routes/Create/index.jsx
@@ -1,7 +1,9 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { Layout } from "../layouts";
-import { createSession } from "../utils/sessionManager";
+import { Layout } from "../../layouts";
+import { createSession } from "../../utils/sessionManager";
+
+import styles from "./index.module.css";
 
 /**
  * Create session component. 
@@ -38,14 +40,14 @@ export class Create extends React.Component {
     return (
       <Layout>
         {Boolean(this.state.activeSession) && (
-          <div style={{ textAlign: "center", margin: 5 }}>
+          <div className={`textAlignCenter ${styles.previousSessions}`}>
             Your Active Session:
             <div>
               <Link
                 to={`/session/${this.state.activeSession}`}
-                style={{ color: "unset" }}
+              // TODO: better link styling, more button-like?
               >
-                {this.state.activeSession}
+                Resume {this.state.activeSession}
               </Link>
             </div>
           </div>
