+++ src/routes/Create/index.module.css
@@ -1,0 +1,3 @@
+.previousSessions {
+    margin: 5px;
+}
