+++ src/routes/FourOhFour/index.jsx
@@ -1,6 +1,8 @@
 import React from "react";
-import { Layout } from "../layouts";
+import { Layout } from "../../layouts";
 import { Link } from "react-router-dom";
+
+// import styles from "./index.module.css";
 
 /**
  * 404 component
@@ -8,8 +10,8 @@ import { Link } from "react-router-dom";
 export const FourOhFour = () => {
   return (
     <Layout>
-      <div style={{ textAlign: "center" }}>
-        <h1 style={{ margin: 0 }}>
+      <div className="textAlignCenter">
+        <h1 className="noMargin">
           <div>404</div>
         </h1>
         <div>What you are looking for ain't here</div>
