+++ src/routes/Room/index.module.css
@@ -1,0 +1,7 @@
+.roomWrapper {
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    flex-wrap: wrap;
+    height: 100%;
+}
