+++ src/routes/index.js
@@ -1,5 +1,5 @@
-export * from "./AuthRedirect";
 export * from "./ConnectToSession";
+export * from "./AuthRedirect";
 export * from "./Create";
 export * from "./Home";
 export * from "./Room";
